feat: integrate pending gameplay and infrastructure updates
feat: add the pending help ticket, dialogue, scripting, plugin, persistence, and runtime session changes already present in the worktree
feat: carry the current combat, loot, potion, clock, bank, mobile, and character service updates together with the related loader and bootstrap changes
feat: include the active network and encryption pipeline updates, client type handling, and packet or handler adjustments already staged in this repository state
refactor: apply the current repository-wide formatting and structural cleanup across server, persistence, networking, scripting, and test projects
refactor: keep solution, project, asset, and configuration files aligned with the formatted codebase state
test: include the synchronized test updates across network packets, persistence, scripting, loaders, handlers, modules, services, plugins, and entity coverage
docs: include the README adjustments already present in the working tree

+++ src/Moongate.Server/Modules/ClockModule.cs
@@ -36,7 +36,12 @@ public sealed class ClockModule
                                   ? spatialConfig.LightSecondsPerUoMinute
                                   : DefaultSecondsPerUoMinute;
 
-        if (!DateTime.TryParse(spatialConfig?.LightWorldStartUtc, CultureInfo.InvariantCulture, DateTimeStyles.AssumeUniversal, out var parsedWorldStart))
+        if (!DateTime.TryParse(
+                spatialConfig?.LightWorldStartUtc,
+                CultureInfo.InvariantCulture,
+                DateTimeStyles.AssumeUniversal,
+                out var parsedWorldStart
+            ))
         {
             _worldStartUtc = DefaultWorldStartUtc;
 
@@ -130,7 +135,7 @@ public sealed class ClockModule
         }
 
         if (_spatialWorldService is not null &&
-            MobileScriptResolver.TryResolveMobile(_spatialWorldService, (uint)mobileId.Value, out mobile))
+            MobileScriptResolver.TryResolveMobile(_spatialWorldService, mobileId.Value, out mobile))
         {
             return true;
         }
